Data-fetching component for /favourite route
